requirements.txt
# LangChain Core Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5

# Document Processing
PyMuPDF==1.23.14
pymupdf==1.23.14

# Vector Store
faiss-cpu==1.7.4

# UI Framework
gradio==4.15.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# Development
pytest==7.4.3
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Additional dependencies for LangChain
tiktoken==0.5.2
openai==1.6.1
